pyproject.toml
[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
authors = [
    {name = "djkcyl", email = "cyl@cyllive.cn"},
]
requires-python = ">=3.9,<4.0"
dependencies = [
    "dnspython>=2.3.0",
    "graiax-silkcoder>=0.3.6",
    "httpx>=0.24.1",
    "peewee>=3.16.2",
    "psutil>=5.9.5",
    "pyyaml>=6.0",
    "pyzbar>=0.1.9",
    "qrcode>=7.4.2",
    "rsa>=4.9",
    "tencentcloud-sdk-python>=3.0.921",
    "wordcloud>=1.9.2",
    "xmltodict>=0.13.0",
    "graia-ariadne[graia]>=0.11.5",
    "opencv-python>=4.7.0.72",
    "pycryptodome>=3.18.0",
    "prompt-toolkit>=3.0.38",
    "toml>=0.10.2",
    "pillow>=9.5.0",
    "motor>=3.2.0",
    "richuru>=0.1.1",
    "beanie>=1.19.2",
    "sentry-sdk>=1.26.0",
    "graiax-playwright>=0.2.4",
    "graiax-text2img-playwright>=0.4.1",
    "hanlp>=2.1.0b50",
    "cnocr>=2.2.2.3",
    "jionlp @ git+https://github.com/dongrixinyu/JioNLP",
    "miniopy-async>=1.15",
    "python-dateutil>=2.8.2",
    "pyecharts>=2.0.3",
]
description = ""
name = "abot-mahv2"
version = "3.0.0"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.ruff]
line-length = 120
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
    "devtools>=0.11.0",
    "flake8>=6.0.0",
    "aio-mc-rcon>=3.2.0",
    "ipython>=8.14.0",
    "ruff>=0.0.277",
]

[tool.pdm.build]
includes = []

